+++ src/typeguards.ts
@@ -15,7 +15,13 @@ export function isDocument<T, S extends RefType>(doc: Ref<T, S>): doc is Documen
  * Check if the given array is already populated
  * @param docs The Array of Refs with uncertain type
  */
-export function isDocumentArray<T, S extends RefType>(docs: Ref<T, S>[]): docs is DocumentType<NonNullable<T>>[] {
+export function isDocumentArray<T, S extends RefType>(
+  docs: mongoose.Types.Array<Ref<T, S>>
+): docs is mongoose.Types.Array<DocumentType<NonNullable<T>>>;
+export function isDocumentArray<T, S extends RefType>(
+  docs: Ref<T, S>[]
+): docs is DocumentType<NonNullable<T>>[];
+export function isDocumentArray(docs: Ref<any, any>[]): unknown { // its "any" & "unkown" because this is not listed as an overload
   return Array.isArray(docs) && docs.every((v) => isDocument(v));
 }
 
@@ -31,7 +37,9 @@ export function isRefType<T, S extends RefType>(doc: Ref<T, S>): doc is NonNulla
  * Check if the document is not undefined/null and is not an document
  * @param docs The Ref with uncretain type
  */
-export function isRefTypeArray<T, S extends RefType>(docs: Ref<T, S>[]): docs is NonNullable<S>[] {
+export function isRefTypeArray<T, S extends RefType>(docs: mongoose.Types.Array<Ref<T, S>>): docs is mongoose.Types.Array<NonNullable<S>>;
+export function isRefTypeArray<T, S extends RefType>(docs: Ref<T, S>[]): docs is NonNullable<S>[];
+export function isRefTypeArray(docs: Ref<any, any>[]): unknown { // its "any" & "unkown" because this is not listed as an overload
   return Array.isArray(docs) && docs.every((v) => isRefType(v));
 }